Find DD4hep before Gaudi to avoid python version clashes
DD4hep has to be found pretty much first for CMake because it is the one
that requires an exact version of python. However, if anything
else (e.g. Boost via Gaudi) finds another version of python (typically a
newer one from the underlying system) then this version will be cached
and the check in the DD4hep cmake configuration will fail.